+++ spec/integrations/before_all_spec.rb
@@ -6,6 +6,6 @@ describe "BeforeAll" do
   specify "it works" do
     output = run_rspec('before_all')
 
-    expect(output).to include("3 examples, 0 failures")
+    expect(output).to include("8 examples, 0 failures")
   end
 end